HDDS-11278. Move code out of Hadoop util package
