+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/ozone/util/CacheMetrics.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.util;
+package org.apache.hadoop.ozone.util;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheStats;
