+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/snapshot/PersistentList.java
@@ -18,7 +18,7 @@
 
 package org.apache.hadoop.ozone.om.snapshot;
 
-import org.apache.hadoop.util.ClosableIterator;
+import org.apache.hadoop.ozone.util.ClosableIterator;
 
 /**
  * Define an interface for persistent list.
